+++ src/TreeList/DefaultItem.tsx
@@ -1,27 +1,19 @@
-import React, { CSSProperties, useMemo } from 'react'
+import React, { useMemo } from 'react'
 import styled from 'styled-components'
 
 import { TreeListItem } from '../useExpandable'
-import { SelectListData } from '../SelectList/SelectList'
 import { ListItem } from '../ListItem'
 import { BranchSpacer } from '../BranchSpacer'
 import { fromTheme } from '../utils'
 import { Icon } from '../Icon'
 import { MdRemove, MdAdd } from 'react-icons/md'
 import { CHILDREN } from '../utils/makeParentChildTree'
+import { ListChildComponentProps } from 'react-window'
 
 export const EXPAND_ICON = MdAdd
 export const COLLAPSE_ICON = MdRemove
 
-const StyledItem = styled.div`
-  li:focus, :hover {
-    button {
-      outline: 5px solid ${fromTheme(theme => theme.colors.lightGray)};
-    }
-  }
-`
-
-const ToggleExpandedButton = styled.button`
+export const ToggleExpandedButton = styled.button`
   width: 20px;
   height: 20px;
   border-radius: 0;
@@ -29,48 +21,65 @@ const ToggleExpandedButton = styled.button`
   border: none;
 
   padding: 0;
+  cursor: pointer;
+`
+
+const StyledListItem = styled(ListItem)`
+  &:focus,
+  &:hover {
+    ${ToggleExpandedButton} {
+      outline: 5px solid ${fromTheme(theme => theme.colors.lightGray)};
+    }
+  }
+`
+
+const StyledValue = styled.span`
+  margin-left: 10px;
 `
 
 const ButtonIcon = styled(Icon)`
   svg {
     width: 100%;
     height: 100%;
-    color: ${fromTheme(theme => theme.colors.gray)}
+    color: ${fromTheme(theme => theme.colors.gray)};
   }
 `
 
-export interface TreeListItemProps<Data extends TreeListItem<SelectListData>> {
-  data: { items: Data[], onChange: (item: Data) => void }
-  style: CSSProperties
-  index: number
-  isScrolling: boolean
+export interface TreeListItemProps<Data extends {}> extends ListChildComponentProps {
+  data: {
+    items: TreeListItem<Data>[]
+    onChange: (item: Data, e?: React.MouseEvent<HTMLLIElement>) => void
+  }
 }
 
-export function DefaultItem<
-  Data extends TreeListItem<SelectListData>
-> ({ data, style, index }: TreeListItemProps<Data>) {
+export function DefaultItem<Data extends { value: string }> ({
+  data,
+  style,
+  index,
+  ...rest
+}: TreeListItemProps<Data>) {
   const { items, onChange } = data
   const item = items[index]
 
-  const handleChange = () => onChange && onChange(item)
-
   return useMemo(() => {
+    const handleChange = (e: React.MouseEvent<HTMLLIElement>) => onChange && onChange(item, e)
+
     return (
-      <StyledItem>
-        <ListItem style={style} onClick={handleChange}>
-          <BranchSpacer
-            spacing={20}
-            index={index}
-            data={items}>
-            {item[CHILDREN] && (
-              <ToggleExpandedButton onClick={item.expand}>
-                <ButtonIcon >{item.expanded ? <COLLAPSE_ICON /> : <EXPAND_ICON />}</ButtonIcon>
-              </ToggleExpandedButton>
-            )}
-            {item.value}
-          </BranchSpacer>
-        </ListItem>
-      </StyledItem>
+      <StyledListItem style={style} onClick={handleChange} {...rest}>
+        <BranchSpacer spacing={20} index={index} data={items}>
+          {item[CHILDREN] && (
+            <ToggleExpandedButton
+              onClick={e => {
+                e.stopPropagation()
+                item.expand()
+              }}
+            >
+              <ButtonIcon>{item.expanded ? <COLLAPSE_ICON /> : <EXPAND_ICON />}</ButtonIcon>
+            </ToggleExpandedButton>
+          )}
+          <StyledValue>{item.value}</StyledValue>
+        </BranchSpacer>
+      </StyledListItem>
     )
-  }, [item, index, items])
+  }, [item, index, items, onChange])
 }