+++ src/TreeList/TreeList.stories.tsx
@@ -4,8 +4,24 @@ import { storiesOf } from '@storybook/react'
 import { TreeList } from './TreeList'
 import { FLAT_LIST_TEST_DATA } from './stories/seed'
 import { FakeRecord } from '../useExpandable/stories/seed'
+import { CustomItem, ModifiedDefaultItem } from './stories/Items'
 
 const parentKeyAccessor = (item: FakeRecord) => item.parentId
+const keyAccessor = (item: FakeRecord) => item.key
 
 storiesOf('organisms|TreeList', module)
-  .add('default', () => <TreeList data={FLAT_LIST_TEST_DATA} parentKeyAccessor={parentKeyAccessor} />)
+  .add('default', () => <TreeList data={FLAT_LIST_TEST_DATA} keyAccessor={keyAccessor} parentKeyAccessor={parentKeyAccessor} />)
+  .add('custom item', () => {
+    return (
+      <TreeList data={FLAT_LIST_TEST_DATA} keyAccessor={keyAccessor} parentKeyAccessor={parentKeyAccessor}>
+        {CustomItem}
+      </TreeList>
+    )
+  })
+  .add('modified item', () => {
+    return (
+      <TreeList data={FLAT_LIST_TEST_DATA} keyAccessor={keyAccessor} parentKeyAccessor={parentKeyAccessor}>
+        {ModifiedDefaultItem}
+      </TreeList>
+    )
+  })