+++ src/TreeList/TreeList.tsx
@@ -1,29 +1,80 @@
-import React, { ReactElement } from 'react'
+import React, { FC, useMemo, ComponentType, useCallback } from 'react'
+import { ReactElementType } from 'react-window'
 
-import { SelectList } from '../SelectList'
-import { SelectListProps, SelectListData } from '../SelectList/SelectList'
-import { useExpandable, TreeListItem } from '../useExpandable'
+import { calculateHeight, addKey } from './helpers'
 import { DefaultItem, TreeListItemProps } from './DefaultItem'
 
-interface TreeListProps<Data extends SelectListData> {
-  parentKeyAccessor: (item: Data) => string
+import { useExpandable } from '../useExpandable'
+import { ListFocus } from '../ListFocus'
+import { StyledList } from '../SelectList'
+
+const MAX_ITEMS_RENDERED = 10
+const ITEM_HEIGHT = 30
+
+interface TreeListProps<Data extends {}> {
+  keyAccessor: (item: Data) => string
+
   data: Data[]
+  parentKeyAccessor: (item: Data) => string | null
   expandedKeys?: string[]
-  children?: ReactElement<TreeListItemProps<TreeListItem<Data>>>
+
+  children?: ComponentType<TreeListItemProps<Data>>
+
+  value?: Data
+  onChange?: (item: Data, e: React.MouseEvent<HTMLLIElement>) => void
+
+  itemHeight?: number
+  maxItemsRendered?: number
+  innerElementType?: ReactElementType
+  width?: string | number
 }
 
-export function TreeList<Data extends SelectListData> ({
+export function TreeList<Data extends {}> ({
   data,
   parentKeyAccessor,
   expandedKeys,
   children,
-  ...rest
-}: SelectListProps & TreeListProps<Data>) {
-  const [expandableData] = useExpandable(data, parentKeyAccessor, expandedKeys)
+  keyAccessor,
+  maxItemsRendered = MAX_ITEMS_RENDERED,
+  itemHeight = ITEM_HEIGHT,
+  onChange,
+  value,
+  innerElementType = 'ul',
+  width = '100%'
+}: TreeListProps<Data>) {
+  const withKeys = useMemo(() => data.map(item => addKey(item, keyAccessor)), [data, keyAccessor])
+  const [expandableData] = useExpandable(withKeys, parentKeyAccessor, expandedKeys)
+
+  const height = calculateHeight(expandableData, itemHeight, maxItemsRendered)
+
+  const handleChange = useCallback((
+    newItem: Data & { key: string },
+    e: React.MouseEvent<HTMLLIElement>
+  ) => {
+    onChange && onChange(data.find(item => keyAccessor(item) === newItem.key)!, e)
+  }, [onChange, data, keyAccessor])
+
+  const itemData = useMemo(() => {
+    return {
+      items: expandableData,
+      onChange: handleChange,
+      selectedItem: value
+    }
+  }, [expandableData, value, handleChange])
 
   return (
-    <SelectList data={expandableData} {...rest}>
-      {children || DefaultItem}
-    </SelectList>
+    <ListFocus>
+      <StyledList
+        height={height}
+        itemData={itemData}
+        itemCount={expandableData.length}
+        itemSize={itemHeight}
+        innerElementType={innerElementType}
+        width={width}>
+        {children || DefaultItem}
+      </StyledList>
+    </ListFocus>
   )
 }
+
+export { ToggleExpandedButton } from './DefaultItem'