+++ src/useSnapEdgeToParent/useSnapEdgeToParent.stories.tsx
@@ -1,4 +1,4 @@
-import React, { FC, useEffect, useMemo } from 'react'
+import React, { FC, useMemo } from 'react'
 
 import { storiesOf } from '@storybook/react'
 
@@ -26,15 +26,11 @@ interface DropdownProps {
 }
 
 const Dropdown: FC<DropdownProps> = ({ breakDistances, initialPositioning, relativeToContainer }) => {
-  const [{ parent, target, container }, positioning, handlePositioning] = useSnapEdgeToParent(breakDistances, initialPositioning)
+  const [{ parent, target, container }, positioning] = useSnapEdgeToParent(breakDistances, initialPositioning)
+
   const position = useMemo(() => {
     return { [positioning.horizontal]: 0, [positioning.vertical]: 20 }
   }, [positioning])
-
-  useEffect(() => {
-    window.addEventListener('resize', handlePositioning)
-    return () => window.removeEventListener('resize', handlePositioning)
-  })
 
   return (
     <Flex